test(meshtastic): migrate utility tests to pytest and split message routing
Convert `TestUncoveredMeshtasticUtilsPaths` from `unittest.TestCase` to a pytest-style class by replacing `self.assert*` methods with native `assert` statements and removing the `unittest` dependency.

Refactor `test_meshtastic_utils_messages.py` by extracting packet routing, portnum classification, and plugin dispatch tests into a new dedicated file `test_meshtastic_utils_messages_routing.py` to improve modularity and reduce file size.
